refactor(buildrequest): extract path-reason classification and metadata plans into metadata.go
Move Concerns D + E + F from orchestrator.go to a new metadata.go
(~720 LOC). Verbatim move with no behaviour change:

- Concern D: BuildRequestAPI* constants, the 11-string PathReason*
  constants, and the ProviderResolution type.
- Concern E: ResolveProviderWithReason.
- Concern F: BuildSingleProviderPlan, BuildSingleProviderPlanForClient,
  BuildFallbackChainPlan, BuildFallbackChainPlanForClient,
  BuildLegacyMetadataPlan, BuildLegacyMetadataPlanForClient.

Drops the now-unused buildrequest/roundrobin import from orchestrator.go
(only metadata.go's plan builders / classifier still call it).
Regenerates bamlutils/embed.go.

Part of #199 item 4.
